fixed events

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/controller/EventController.java
@@ -8,6 +8,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("events")
-public class EventController extends BaseReadWriteController<EventPayload, Event, Long> {
+public class EventController extends BaseReadWriteController<EventPayload, Event, Integer> {
 
 }
